==> R/00-setup_packages.R <==
# This code installs the packages only available on GitHub (not on CRAN)
renv::install(c(
  "Tom-Wolff/chiSTIGmodules@main",
  "EpiModel/EpiModelHIV-p",
 "EpiModel/EpiModelHPC"
))

# This code finds and install the libraries used by the project (CRAN version)
renv::hydrate()

# Force `renv` to discover the following packages
if (FALSE) {
  library("rmarkdown")
  library("pkgload")
  library("sessioninfo")
}
